# 4.x: Improve Return/Throw/Append/Prepend with ImmediateScheduler