@danielcweber 2018-07-04
I'm very fine with inlining it into `Return` and `Throw` because they are very basic, and would merge it right away, but for `Append` and `Prepend`, I really don't know. We want to lower the barriers for writing new operators, this is just a bad sign for new contributors if it would now take separate classes for special schedulers. Also, there is so much code duplication in AppendPrepend. Can't `ScheduleAction` do the check on `Immediate`? The performance penalty is ok, it would just be a reference comparison.