@danielcweber 2018-07-04
Or at least, see whether the duplicated code can be pushed back into a base class.