@akarnokd 2018-07-04
> Can't ScheduleAction do the check on Immediate?

That's a run-time check whereas this does this check at assembly time and picks the optimized version, which then on remains the same for all further subscriptions.

> Can't ScheduleAction do the check on Immediate? The performance penalty is ok, it would just be a reference comparison.

It's not just the check, but any other code & field that can be eliminated in a dedicated implementation.

> We want to lower the barriers for writing new operators, this is just a bad sign for new contributors if it would now take separate classes for special schedulers.

Writing new operators is one thing, optimizing them is another.